class-35/demo/pyproject.toml
[tool.poetry]
name = "python"
version = "0.1.0"
description = ""
authors = ["JB-Tellez <jb.tellez@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pytest-watch = "^4.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
